feat: Consolidate UI DSL and Style definitions into a single file for improved organization
